Move io sub-package and spect module into vak/prep/spectrogram_dataset/
