+++ src/vak/__init__.py
@@ -18,7 +18,6 @@ from . import (
     curvefit,
     datasets,
     eval,
-    io,
     learncurve,
     metrics,
     models,
@@ -27,7 +26,6 @@ from . import (
     plot,
     predict,
     prep,
-    spect,
     split,
     train,
     transforms,
@@ -40,7 +38,6 @@ __all__ = [
     "common",
     "config",
     "datasets",
-    "io",
     "learncurve",
     "metrics",
     "models",
@@ -48,7 +45,6 @@ __all__ = [
     "plot",
     "predict",
     "prep",
-    "spect",
     "split",
     "train",
     "transforms",
