--- src/vak/io/__init__.py
@@ -1,5 +1,0 @@
-"""module that handles file input-output:
-- audio files
-- spectrograms made from audio files of vocalizations
-- .csv files that represent a dataset of vocalizations that combines all those files together"""
-from . import audio, dataframe, spect
