+++ src/vak/prep/__init__.py
@@ -1,10 +1,12 @@
 from . import (
-    prep_helper
+    prep_helper,
+    spectrogram_dataset,
 )
 from .prep import *
 
 
 __all__ = [
     'prep',
-    'prep_helper'
+    'prep_helper',
+    'spectrogram_dataset',
 ]
