+++ src/vak/prep/spectrogram_dataset/__init__.py
@@ -1,0 +1,1 @@
+from .prep import prep_spectrogram_dataset
